Reorganize components into UI and feature
