+++ src/App.tsx
@@ -2,17 +2,17 @@ import { Analytics } from '@vercel/analytics/react';
 import { AnimatePresence } from 'framer-motion';
 import React, { useEffect, useState } from 'react';
 
-import Button from './components/Button';
 import FeedbackButton from './components/FeedbackButton';
 import MovieDeck from './components/MovieDeck';
 import MovieFilters from './components/MovieFilters';
-import MovieModal from './components/MovieModal';
 import MoviePicker from './components/MoviePicker';
 import MyList from './components/MyList';
-import SectionLabel from './components/SectionLabel';
 import SupportButton from './components/SupportButton';
 import TMDBAttribution from './components/TMDBAttribution';
 import TonightsPick from './components/TonightsPick';
+import Button from './components/ui/Button';
+import MovieModal from './components/ui/MovieModal';
+import SectionLabel from './components/ui/SectionLabel';
 import type { FilterOptions, Movie } from './types';
 import { exportWatchlistCSV, filterMovies } from './utils';
 import { enrichAllMovies } from './utils/tmdb';
@@ -164,8 +164,9 @@ const App: React.FC = () => {
           <h1
             style={{
               fontFamily: 'var(--font-display)',
-              fontSize: '2.5rem',
-              fontWeight: 800,
+              fontSize: 'var(--text-display)',
+              fontWeight:
+                'var(--weight-display)' as React.CSSProperties['fontWeight'],
               color: 'var(--color-text)',
               lineHeight: 1.05,
               letterSpacing: '-0.04em',
@@ -175,10 +176,11 @@ const App: React.FC = () => {
           </h1>
           <p
             style={{
-              fontSize: '0.9rem',
+              fontSize: 'var(--text-base)',
               color: 'var(--color-text-secondary)',
               marginTop: '6px',
-              fontWeight: 500,
+              fontWeight:
+                'var(--weight-medium)' as React.CSSProperties['fontWeight'],
             }}
           >
             From your watchlist to tonight's pick
@@ -198,15 +200,16 @@ const App: React.FC = () => {
               {movies.length > 0 && (
                 <span
                   style={{
-                    fontSize: '0.7rem',
-                    fontWeight: 700,
+                    fontSize: 'var(--text-xs)',
+                    fontWeight:
+                      'var(--weight-bold)' as React.CSSProperties['fontWeight'],
                     color:
                       activeTab === 'list'
                         ? 'var(--color-accent)'
                         : 'var(--color-muted)',
                     background: 'var(--color-surface-2)',
                     border: '1px solid var(--color-border)',
-                    borderRadius: '10px',
+                    borderRadius: 'var(--radius-pill)',
                     padding: '1px 7px',
                     transition: 'color 0.15s',
                   }}
@@ -262,13 +265,18 @@ const App: React.FC = () => {
                   padding: '60px 0',
                 }}
               >
-                <span style={{ fontSize: '2.5rem', opacity: 0.35 }}>🎬</span>
+                <span
+                  style={{ fontSize: 'var(--text-display)', opacity: 0.35 }}
+                >
+                  🎬
+                </span>
 
                 <div>
                   <p
                     style={{
-                      fontSize: '1rem',
-                      fontWeight: 700,
+                      fontSize: 'var(--text-md)',
+                      fontWeight:
+                        'var(--weight-bold)' as React.CSSProperties['fontWeight'],
                       color: 'var(--color-text)',
                     }}
                   >
@@ -277,9 +285,10 @@ const App: React.FC = () => {
 
                   <p
                     style={{
-                      fontSize: '0.875rem',
+                      fontSize: 'var(--text-base)',
                       color: 'var(--color-text-secondary)',
-                      fontWeight: 500,
+                      fontWeight:
+                        'var(--weight-medium)' as React.CSSProperties['fontWeight'],
                       marginTop: '4px',
                     }}
                   >
@@ -311,15 +320,17 @@ const App: React.FC = () => {
                     <SectionLabel noMargin>Filters</SectionLabel>
                     <span
                       style={{
-                        fontSize: '0.875rem',
+                        fontSize: 'var(--text-base)',
                         color: 'var(--color-text-secondary)',
-                        fontWeight: 500,
+                        fontWeight:
+                          'var(--weight-medium)' as React.CSSProperties['fontWeight'],
                       }}
                     >
                       <span
                         style={{
                           color: 'var(--color-text)',
-                          fontWeight: 700,
+                          fontWeight:
+                            'var(--weight-bold)' as React.CSSProperties['fontWeight'],
                         }}
                       >
                         {filteredMovies.length}
@@ -443,9 +454,10 @@ const App: React.FC = () => {
         <div style={{ display: 'flex', alignItems: 'center', gap: '12px' }}>
           <p
             style={{
-              fontSize: '0.8rem',
+              fontSize: 'var(--text-sm)',
               color: 'var(--color-muted)',
-              fontWeight: 500,
+              fontWeight:
+                'var(--weight-medium)' as React.CSSProperties['fontWeight'],
             }}
           >
             © 2026 CueMovie · v{APP_VERSION}
@@ -485,8 +497,10 @@ const TabButton: React.FC<{
         ? '2px solid var(--color-accent)'
         : '2px solid transparent',
       color: active ? 'var(--color-text)' : 'var(--color-muted)',
-      fontSize: '0.875rem',
-      fontWeight: active ? 700 : 500,
+      fontSize: 'var(--text-base)',
+      fontWeight: active
+        ? ('var(--weight-bold)' as React.CSSProperties['fontWeight'])
+        : ('var(--weight-medium)' as React.CSSProperties['fontWeight']),
       fontFamily: 'var(--font-body)',
       cursor: 'pointer',
       transition: 'color 0.15s',
@@ -518,8 +532,8 @@ const DeckToggle: React.FC<{ enabled: boolean; onToggle: () => void }> = ({
       border: 'none',
       background: 'transparent',
       color: enabled ? 'var(--color-accent)' : 'var(--color-muted)',
-      fontSize: '0.8rem',
-      fontWeight: 700,
+      fontSize: 'var(--text-sm)',
+      fontWeight: 'var(--weight-bold)' as React.CSSProperties['fontWeight'],
       fontFamily: 'var(--font-body)',
       cursor: 'pointer',
       letterSpacing: '0.05em',
