+++ src/components/ui/MovieCard.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import type { Movie } from '../types';
+import type { Movie } from '../../types';
 import GenrePill from './GenrePill';
 
 type Props = {
